separate debug monitoring client
Added a rover_status script which receives a debugging object from
driver.py via socket. This allows state examination to be displayed in a
separate tmux pane alongside an AI playthrough. Further refinements
could be made to include more information and even allow for
interactively exploring the contents of the knowledge base.